Feat: Add drop down list and update navigation links
Updated navigation buttons so that they now consist of 'about', 'resume', and 'contact'. The 'about' button I hope to add somekind of functionality to display information about me. The 'resume' button links out to my resume site. The 'contact' button is a Bootstrap drop down list, which lists links for my LinkedIn & GitHub profiles, and a mailto link for sending me email.

Still need to update css so that the new 'contact' button is consistent with the rest, but I plan to change the entire site style anyways.